requirements.txt
coverage==7.4.4
crispy-bootstrap5==2024.2
Django==5.0.3
django-crispy-forms==2.1
django-debug-toolbar==4.3.0
django-environ==0.11.2
django-extensions==3.2.3
django-filter==24.1
django-silk==5.1.0
djangorestframework==3.15.1
factory-boy==3.3.0
Markdown==3.6
parameterized==0.9.0
pre-commit==3.7.0
psycopg2==2.9.9
drf-yasg==1.21.7
